Fix savedInstanceState errors when using NavigationComponent
When using Google's Navigation component, onDestroyView is called when navigating away from a screen.
This means that when onSaveInstanceState is called, the view isn't available to for us to save the sate from.

This commit saves state into a temp bundle each time the Fragment's view is destroyed and then uses
that temp bundle when onSaveInstanceState is actually called.